refactor: 커서 페이징 DTO 위치 이동
