+++ techeerzip/src/main/java/backend/techeerzip/domain/session/controller/SessionController.java
@@ -3,10 +3,10 @@ package backend.techeerzip.domain.session.controller;
 import backend.techeerzip.domain.session.dto.request.SessionCreateRequest;
 import backend.techeerzip.domain.session.dto.response.SessionResponse;
 import backend.techeerzip.domain.session.service.SessionService;
-import backend.techeerzip.global.common.CursorPageViewCountRequest;
-import backend.techeerzip.global.common.CursorPageViewCountResponse;
-import backend.techeerzip.global.common.CursorPageCreatedAtRequest;
-import backend.techeerzip.global.common.CursorPageCreatedAtResponse;
+import backend.techeerzip.domain.session.dto.request.CursorPageViewCountRequest;
+import backend.techeerzip.domain.session.dto.response.CursorPageViewCountResponse;
+import backend.techeerzip.domain.session.dto.request.CursorPageCreatedAtRequest;
+import backend.techeerzip.domain.session.dto.response.CursorPageCreatedAtResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import org.springdoc.core.annotations.ParameterObject;
