+++ techeerzip/src/main/java/backend/techeerzip/domain/session/service/SessionService.java
@@ -8,10 +8,10 @@ import backend.techeerzip.domain.session.exception.SessionUnauthorizedException;
 import backend.techeerzip.domain.session.repository.SessionRepository;
 import backend.techeerzip.domain.user.entity.User;
 import backend.techeerzip.domain.user.repository.UserRepository;
-import backend.techeerzip.global.common.CursorPageViewCountRequest;
-import backend.techeerzip.global.common.CursorPageViewCountResponse;
-import backend.techeerzip.global.common.CursorPageCreatedAtRequest;
-import backend.techeerzip.global.common.CursorPageCreatedAtResponse;
+import backend.techeerzip.domain.session.dto.request.CursorPageViewCountRequest;
+import backend.techeerzip.domain.session.dto.response.CursorPageViewCountResponse;
+import backend.techeerzip.domain.session.dto.request.CursorPageCreatedAtRequest;
+import backend.techeerzip.domain.session.dto.response.CursorPageCreatedAtResponse;
 import backend.techeerzip.global.exception.ErrorCode;
 import backend.techeerzip.global.exception.NotFoundException;
 import org.springframework.stereotype.Service;
